Find a single user using jpa repository

+++ src/main/java/com/springbootrestful/restfulapis/users/User.java
@@ -11,7 +11,7 @@ import java.util.Date;
 public class User {
     @Id
     @GeneratedValue
-    private Integer id;
+    private Long id;
 
     @Size(min = 2, message = "Name should have at least two characters")
     private String name;
@@ -21,17 +21,17 @@ public class User {
 
     protected User() {}
 
-    public User(Integer id, String name, Date birthDate) {
+    public User(Long id, String name, Date birthDate) {
         this.id = id;
         this.name = name;
         this.birthDate = birthDate;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
